+++ diagram-viewer/src/main/java/com/powsybl/diagram/viewer/nad/NetworkAreaDiagramController.java
@@ -6,7 +6,7 @@
  */
 package com.powsybl.diagram.viewer.nad;
 
-import com.google.common.io.ByteStreams;
+import com.powsybl.diagram.viewer.common.AbstractDiagramController;
 import com.powsybl.iidm.network.Container;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.Substation;
@@ -17,9 +17,6 @@ import javafx.concurrent.Service;
 import javafx.concurrent.Task;
 import javafx.concurrent.Worker;
 import javafx.fxml.FXML;
-import javafx.scene.control.TextArea;
-import javafx.scene.input.ScrollEvent;
-import javafx.scene.web.WebView;
 import netscape.javascript.JSObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -27,51 +24,23 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 /**
  * @author Florian Dupuy <florian.dupuy at rte-france.com>
  */
-public class NetworkAreaDiagramController {
+public class NetworkAreaDiagramController extends AbstractDiagramController {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(NetworkAreaDiagramController.class);
 
     @FXML
-    public WebView diagramWebView;
-    @FXML
-    public TextArea svgContent;
-    @FXML
-    public TextArea info;
-
-    private String html;
-    private String js;
-
-    private final NetworkAreaDiagramJsHandler jsHandler = new NetworkAreaDiagramJsHandler();
-
-    @FXML
     private void initialize() throws IOException {
-        // Add Zoom management
-        diagramWebView.addEventFilter(ScrollEvent.SCROLL, (ScrollEvent e) -> {
-            if (e.isControlDown()) {
-                double deltaY = e.getDeltaY();
-                double zoom = diagramWebView.getZoom();
-                if (deltaY < 0) {
-                    zoom /= 1.1;
-                } else if (deltaY > 0) {
-                    zoom *= 1.1;
-                }
-                diagramWebView.setZoom(zoom);
-                e.consume();
-            }
-        });
+        super.init();
 
-        // Avoid the useless right click on the image
-        diagramWebView.setContextMenuEnabled(false);
+        setUpListenerOnWebViewChanges(new NetworkAreaDiagramJsHandler());
+    }
 
+    protected void setUpListenerOnWebViewChanges(java.lang.Object jsHandler) {
         // Set up the listener on WebView changes
         // A listener has to be added as loading takes time - execute once the content is successfully loaded
         diagramWebView.getEngine().getLoadWorker().stateProperty().addListener((observable, oldValue, newValue) -> {
@@ -83,17 +52,10 @@ public class NetworkAreaDiagramController {
                         "{ jsHandler.log(message); };");
             }
         });
-
-        html = new String(ByteStreams.toByteArray(Objects.requireNonNull(getClass().getResourceAsStream("/nad/svg.html"))));
-        js = new String(ByteStreams.toByteArray(Objects.requireNonNull(getClass().getResourceAsStream("/nad/svg.js"))));
     }
 
     public void createDiagram(Network network, NetworkAreaDiagramModel model, StringProperty modelSvgContent, Container<?> container) {
-        modelSvgContent.addListener((obs, oldV, newV) -> {
-            String embeddedSvg = html.replaceFirst("%__SVG__%", newV).replaceFirst("%__JS__%", js);
-            diagramWebView.getEngine().loadContent(embeddedSvg);
-        });
-        svgContent.textProperty().bind(modelSvgContent);
+        super.createDiagram(container, modelSvgContent);
 
         updateDiagram(network, model, modelSvgContent, container);
     }
@@ -139,30 +101,5 @@ public class NetworkAreaDiagramController {
             default:
                 throw new AssertionError();
         }
-    }
-
-    public void onClickFitToContent() {
-        String svgData = svgContent.getText();
-        Optional<String> svgLine = svgData.lines().filter(l -> l.contains("<svg")).findAny();
-        if (svgLine.isPresent()) {
-            String valuePattern = "\"([^\"]*)\"";
-            Pattern pW = Pattern.compile("width=" + valuePattern);
-            Matcher mW = pW.matcher(svgLine.get());
-            Pattern pH = Pattern.compile("height=" + valuePattern);
-            Matcher mH = pH.matcher(svgLine.get());
-            if (mH.find() && mW.find()) {
-                double svgWidth = Double.parseDouble(mW.group(1));
-                double svgHeight = Double.parseDouble(mH.group(1));
-                double paneWidth = diagramWebView.widthProperty().get();
-                double paneHeight = diagramWebView.heightProperty().get();
-                double zoomH = paneHeight / svgHeight;
-                double zoomW = paneWidth / svgWidth;
-                diagramWebView.setZoom(Math.min(zoomH, zoomW));
-            }
-        }
-    }
-
-    public void onClickResetZoom() {
-        diagramWebView.setZoom(1.0);
     }
 }
