+++ diagram-viewer/src/main/java/com/powsybl/diagram/viewer/sld/SingleLineDiagramController.java
@@ -6,7 +6,7 @@
  */
 package com.powsybl.diagram.viewer.sld;
 
-import com.google.common.io.ByteStreams;
+import com.powsybl.diagram.viewer.common.AbstractDiagramController;
 import com.powsybl.iidm.network.*;
 import com.powsybl.sld.SingleLineDiagram;
 import com.powsybl.sld.layout.LayoutParameters;
@@ -19,8 +19,6 @@ import javafx.concurrent.Task;
 import javafx.concurrent.Worker;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
-import javafx.scene.input.ScrollEvent;
-import javafx.scene.web.WebView;
 import netscape.javascript.JSObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -28,23 +26,13 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.io.UncheckedIOException;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * @author Thomas Adam <tadam at slicom.fr>
  */
-public class SingleLineDiagramController {
+public class SingleLineDiagramController extends AbstractDiagramController {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(SingleLineDiagramController.class);
-
-    @FXML
-    public WebView diagramWebView;
-
-    @FXML
-    public TextArea svgContent;
 
     @FXML
     public TextArea metadataContent;
@@ -53,44 +41,11 @@ public class SingleLineDiagramController {
     public TextArea graphContent;
 
     @FXML
-    public TextArea info;
-
-    private String html;
-
-    private String js;
-
-    @FXML
     private void initialize() throws IOException {
-        // Add Zoom management
-        diagramWebView.addEventFilter(ScrollEvent.SCROLL, (ScrollEvent e) -> {
-            if (e.isControlDown()) {
-                double deltaY = e.getDeltaY();
-                double zoom = diagramWebView.getZoom();
-                if (deltaY < 0) {
-                    zoom /= 1.1;
-                } else if (deltaY > 0) {
-                    zoom *= 1.1;
-                }
-                diagramWebView.setZoom(zoom);
-                e.consume();
-            }
-        });
-
-        // Avoid the useless right click on the image
-        diagramWebView.setContextMenuEnabled(false);
-
-        html = new String(ByteStreams.toByteArray(Objects.requireNonNull(getClass().getResourceAsStream("/sld/svg.html"))));
-        js = new String(ByteStreams.toByteArray(Objects.requireNonNull(getClass().getResourceAsStream("/sld/svg.js"))));
+        super.init();
     }
 
-    public void createDiagram(SingleLineDiagramJsHandler jsHandler,
-                              Network network,
-                              ReadOnlyBooleanProperty showNamesProperty,
-                              SingleLineDiagramModel model,
-                              SingleLineDiagramModel.ContainerResult containerResult,
-                              Container<?> container) {
-        info.setText(String.join(System.lineSeparator(), "id: " + container.getId(), "name: " + container.getNameOrId()));
-
+    protected void setUpListenerOnWebViewChanges(java.lang.Object jsHandler) {
         // Set up the listener on WebView changes
         // A listener has to be added as loading takes time - execute once the content is successfully loaded
         diagramWebView.getEngine().getLoadWorker().stateProperty().addListener((observable, oldValue, newValue) -> {
@@ -102,7 +57,17 @@ public class SingleLineDiagramController {
                         "{ jsHandler.log(message); };");
             }
         });
+    }
 
+    public void createDiagram(SingleLineDiagramJsHandler jsHandler,
+                              Network network,
+                              ReadOnlyBooleanProperty showNamesProperty,
+                              SingleLineDiagramModel model,
+                              SingleLineDiagramModel.ContainerResult containerResult,
+                              Container<?> container) {
+        super.createDiagram(container, containerResult.svgContentProperty());
+
+        // JSHandler management
         jsHandler.setOperateSwitch(swId -> {
             Switch sw = network.getSwitch(swId);
             if (sw != null) {
@@ -112,15 +77,10 @@ public class SingleLineDiagramController {
                 updateDiagram(network, showNamesProperty, model, containerResult, container);
             }
         });
+        setUpListenerOnWebViewChanges(jsHandler);
+        containerResult.metadataContentProperty().addListener((obs, oldV, newV) -> jsHandler.setMetadata(containerResult.metadataContentProperty().get()));
 
-        containerResult.svgContentProperty().addListener((obs, oldV, newV) -> {
-            String embeddedSvg = html.replaceFirst("%__SVG__%", newV).replaceFirst("%__JS__%", js);
-            diagramWebView.getEngine().loadContent(embeddedSvg);
-        });
-        containerResult.metadataContentProperty().addListener((obs, oldV, newV) -> {
-            jsHandler.setMetadata(containerResult.metadataContentProperty().get());
-        });
-        svgContent.textProperty().bind(containerResult.svgContentProperty());
+        // Metadata & Graph binding
         metadataContent.textProperty().bind(containerResult.metadataContentProperty());
         graphContent.textProperty().bind(containerResult.jsonContentProperty());
 
@@ -181,34 +141,5 @@ public class SingleLineDiagramController {
             LOGGER.error(exception.toString(), exception);
         });
         sldService.start();
-    }
-
-    public void onClickFitToContent() {
-        String svgData = svgContent.getText();
-        Optional<String> svgLine = svgData.lines().filter(l -> l.contains("<svg")).findAny();
-        if (svgLine.isPresent()) {
-            String valuePattern = "\"([^\"]*)\"";
-            Pattern pW = Pattern.compile("width=" + valuePattern);
-            Matcher mW = pW.matcher(svgLine.get());
-            Pattern pH = Pattern.compile("height=" + valuePattern);
-            Matcher mH = pH.matcher(svgLine.get());
-            if (mH.find() && mW.find()) {
-                double svgWidth = Double.parseDouble(mW.group(1));
-                double svgHeight = Double.parseDouble(mH.group(1));
-                double paneWidth = diagramWebView.widthProperty().get();
-                double paneHeight = diagramWebView.heightProperty().get();
-                double zoomH = paneHeight / svgHeight;
-                double zoomW = paneWidth / svgWidth;
-                diagramWebView.setZoom(Math.min(zoomH, zoomW));
-            }
-        }
-    }
-
-    public void onClickResetZoom() {
-        diagramWebView.setZoom(1.0);
-    }
-
-    public void clean() {
-        info.setText("");
     }
 }
